backend-api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.3.1
    # via kombu
asgiref==3.10.0
    # via
    #   django
    #   django-cors-headers
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
billiard==4.2.2
    # via celery
celery==5.5.3
    # via -r requirements.in
certifi==2025.10.5
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
django==5.2.8
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
    #   drf-yasg
django-cors-headers==4.9.0
    # via -r requirements.in
django-extensions==4.1
    # via -r requirements.in
django-filter==25.2
    # via -r requirements.in
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
    #   drf-yasg
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
drf-spectacular==0.29.0
    # via -r requirements.in
drf-yasg==1.21.11
    # via -r requirements.in
et-xmlfile==2.0.0
    # via openpyxl
factory-boy==3.3.3
    # via -r requirements.in
faker==37.12.0
    # via
    #   -r requirements.in
    #   factory-boy
gunicorn==23.0.0
    # via -r requirements.in
idna==3.11
    # via requests
inflection==0.5.1
    # via
    #   drf-spectacular
    #   drf-yasg
jsonschema==4.25.1
    # via drf-spectacular
jsonschema-specifications==2025.9.1
    # via jsonschema
kombu==5.5.4
    # via celery
numpy==2.3.4
    # via pandas
openpyxl==3.1.5
    # via -r requirements.in
packaging==25.0
    # via
    #   drf-yasg
    #   gunicorn
    #   kombu
pandas==2.3.3
    # via -r requirements.in
pillow==12.0.0
    # via -r requirements.in
prompt-toolkit==3.0.52
    # via click-repl
psycopg2==2.9.11
    # via -r requirements.in
psycopg2-binary==2.9.11
    # via -r requirements.in
pyjwt==2.10.1
    # via djangorestframework-simplejwt
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   celery
    #   pandas
python-decouple==3.8
    # via -r requirements.in
python-dotenv==1.2.1
    # via -r requirements.in
pytz==2025.2
    # via
    #   -r requirements.in
    #   drf-yasg
    #   pandas
pyyaml==6.0.3
    # via
    #   drf-spectacular
    #   drf-yasg
redis==7.0.1
    # via -r requirements.in
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via -r requirements.in
rpds-py==0.28.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
typing-extensions==4.15.0
    # via referencing
tzdata==2025.2
    # via
    #   faker
    #   kombu
    #   pandas
uritemplate==4.2.0
    # via
    #   drf-spectacular
    #   drf-yasg
urllib3==2.5.0
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.14
    # via prompt-toolkit
